pyproject.toml
[project]
name = "mef"
version = "0.1.0"
description = "Muse Engine Forecaster — daily forecasting & recommendation tool over a curated US equity/ETF universe"
requires-python = ">=3.12"
dependencies = [
  "psycopg2-binary>=2.9.9",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
mef = "mef.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["mef*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
